* commit '0a9f8061388ffbbf3732e452417506e43926e3b7':
  accCmd: set apache as owner for all billing files even if command is calles as root (via Command line)
  Views: added success and error message on deleting related models
  Settlementrun: show downloadable files in proper Order
  SepaAcc: do not echo paths on accCmd run
  Bugfix-Product: fixed bugs of sales & purchase tariff change, added 2 more decimal places (now 4) for product price
  accCmd: Select Call Data Records indexed by contract number (not id) too
  Modem: Changed Red to Yellow Color for Modems with critical transmit power level
  Bugfix-accCmd: create data directories with accurate permissions
  Bugfix-accCmd: 3 more Error Checkings (2x array index=0, 1x pdflatex installed)
  Bugfix-BaseController: Show error message when no model is specified for Deletion - see TODO\!
  Views: Fixed Color bug on Modem index label, array to string bug on ProvMon view, temporary removing of ccc module bug for contracts
  SettlementRun: changed order of commands in create function to fix create bug, change rights of billing files directory
  Authentication: use Modem Model for ProvMon Auth

+++ app/Http/Controllers/BaseController.php
@@ -512,17 +512,23 @@ class BaseController extends Controller {
 	 */
 	public function destroy($id)
 	{
+		// bulk delete
 		if ($id == 0)
 		{
-			// bulk delete
+			// Error Message when no Model is specified - NOTE: delete_message must be an array of the structure below !
+			if (!isset(Input::all()['ids']))
+				return Redirect::back()->with('delete_message', ['message' => 'No Entry For Deletion specified', 'class' => \NamespaceController::get_route_name(), 'color' => 'red']);
+
 			foreach (Input::all()['ids'] as $id => $val)
 				static::get_model_obj()->findOrFail($id)->delete();
 		}
 		else
 			static::get_model_obj()->findOrFail($id)->delete();
 
-		return Redirect::back();
+		$class = \NamespaceController::get_route_name();
+		return Redirect::back()->with('delete_message', ['message' => 'Successful deleted '.$class, 'class' => $class, 'color' => 'blue']);
 	}
+
 
 
 // Deprecated:
